\name{allele}
\alias{allele}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Retrieve allele from PDEnv for SNP arrays}
\description{
  The function retrieves the vector of alleles from the Platform Design
  Environment for SNP arrays.
}
\usage{
allele(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{an FeatureSet object.}
}
\value{
  A character vector of alleles.
}
\examples{
## DO NOT RUN
## FeatureSetObject <- read.celfiles(list.celfiles())
## theAlleles <- allele(FeatureSetObject)
}
\keyword{manip}% at least one, from doc/KEYWORDS
