\name{characterOrNULL-class}
\docType{class}
\alias{characterOrNULL-class}

\title{Class "characterOrNULL"}
\description{Union of character and NULL}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "characterOrNULL" in the signature.
}
\keyword{classes}
