\name{DBPDInfo-class}
\docType{class}
\alias{DBPDInfo-class}
\alias{AffySNPPDInfo-class}

\title{Class "DBPDInfo"}
\description{A class for Platform Design Information objects, stored
  using a database approach}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("DBPDInfo", ...)}.
}
\section{Slots}{
	 \describe{
    \item{\code{getdb}:}{Object of class \code{"function"}  }
    \item{\code{tableInfo}:}{Object of class \code{"data.frame"}  }
    \item{\code{manufacturer}:}{Object of class \code{"character"}  }
    \item{\code{genomebuild}:}{Object of class \code{"character"}  }
    \item{\code{geometry}:}{Object of class \code{"integer"} with length
    2 (rows x columns)}
  }
}
\section{Extends}{
Class \code{"\linkS4class{PDInfo}"}, directly.
}
\section{Methods}{
No methods defined with class "DBPDInfo" in the signature.
}
\references{}
\author{}
\note{}

\seealso{
}
\examples{
##---- Should be DIRECTLY executable !! ----
}
\keyword{classes}
