\name{featureIndex}
\alias{featureIndex}
\alias{featureIndex,FeatureSet-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Indexes for features in the expression matrix}
\description{
  Indexes for features (PM/MM/both) in the expression matrix.
}
\usage{
featureIndex(object, which = c("both", "pm", "mm"), subset = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{FeatureSet} object}
  \item{which}{subset of interest: "both", "pm" or "mm"?}
  \item{subset}{not implemented yet.}
}
\details{
}
\value{
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}

\keyword{manip}
