\name{featureInfo}
\alias{featureInfo}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Platform Design Environment}
\description{
  Accessor for the Platform Design Environment.
}
\usage{
featureInfo(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{platformDesign} object.}
}
\value{
  Returns the environment containing the Platform Design information.
}
\seealso{\code{\link{getPD}}}
\keyword{manip}% at least one, from doc/KEYWORDS
