\name{Methods for FeatureSet}
\docType{methods}

\alias{platform<-,FeatureSet-method}
\alias{manufacturer,FeatureSet-method}
\alias{manufacturer<-,FeatureSet-method}
\alias{exprs,FeatureSet-method}
\alias{initialize, FeatureSet-method}

\title{Methods for FeatureSet objects}
\description{
  Methods for FeatureSet objects.
}
\section{Methods}{
\describe{
\item{object = "FeatureSet"}{FeatureSet object}
\item{.Object = "FeatureSet"}{Instance of FeatureSet}
}}

\keyword{methods}

