\name{File Manipulation}
\alias{list.celfiles}
\alias{list.xysfiles}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Functions to list CEL and XYS files.}
\description{
  These functions list CEL and XYS files.
}
\usage{
list.celfiles(...)
list.xysfiles(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{arguments to be passed to \code{list.files}.}
}
\seealso{\code{\link{list.files}}}
\keyword{manip}% at least one, from doc/KEYWORDS
