\name{genomeBuild}
\alias{genomeBuild}
\alias{genomeBuild,TilingFeatureSet-method}
\alias{genomeBuild,PDInfo-method}
\alias{genomeBuild,FeatureSet-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Genome Build Information}
\description{
  Returns the genome build information
}
\usage{
genomeBuild(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{TilingFeatureSet} or \code{PDInfo} object.}
}
\details{
}
\value{
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{manip}

