\name{getAffinitySplineCoefficients}
\alias{getAffinitySplineCoefficients}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Estimate affinity coefficients.
}
\description{
  Estimate affinity coefficients using sequence information and splines.
}
\usage{
getAffinitySplineCoefficients(intensities, sequences)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{intensities}{Intensity matrix}
  \item{sequences}{Probe sequences}
}
\value{
  Matrix with estimated coefficients.
}
\seealso{
getBaseProfile
}

\keyword{manip}

