\name{getPlatformDesign}
\alias{getPlatformDesign}
\alias{getPD}
\alias{getPlatformDesign,FeatureSet-method}

\title{Retrieve Platform Design object}
\description{
  Retrieve platform design object.
}
\usage{
getPlatformDesign(object)
getPD(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{FeatureSet} object}
}
\details{
}
\value{
  \code{platformDesign} or \code{PDInfo} object.
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}

\keyword{manip}

