\name{image}
\alias{image}
\alias{image,FeatureSet-method}
\title{Display a Color Image}
\description{
  Produces an image (\code{graphics::image}) for each sample.
}
\usage{
image(x, ...)
}
\arguments{
  \item{x}{\code{FeatureSet} object}
  \item{\dots}{parameters to be passed to \code{plot}}
}
% Add examples
\keyword{hplot}
