\name{mmindex}
\alias{mmindex}
\alias{pmindex}
\alias{mmindex,AffySNPPDInfo-method}
\alias{pmindex,AffySNPPDInfo-method}
\alias{mmindex,FeatureSet-method}
\alias{pmindex,FeatureSet-method}
\alias{mmindex,platformDesign-method}
\alias{pmindex,platformDesign-method}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{Extract PM or MM indexes}
\description{
  Extracts the indexes for PM or MM probes, which can be used with the
  intensity matrix for subsetting.
}
\usage{
mmindex(object)
pmindex(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{AffySNPPDInfo}, \code{FeatureSet} or
    \code{platformDesign} object}
}
\details{
}
\value{
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}

\keyword{manip}
