\name{indexFeatureSetName}
\alias{indexFeatureSetName}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Indexes for feature sets.}
\description{
  The function returns the indexes in the PDEnv for given feature sets.
}
\usage{
indexFeatureSetName(object, featurenames)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{An \code{FeatureSet} object.}
  \item{featurenames}{A character vector with the feature set names.}
}
\keyword{manip}% at least one, from doc/KEYWORDS
