\name{i2xy}
\alias{i2xy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Index <-> XY}
\description{
  Converts index to XY coordinates and vice-versa.
}
\usage{
i2xy(i, obatch)
xy2i(x,y,obatch)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{i}{vector of indexes}
  \item{x}{vector of X locations}
  \item{y}{vector of Y locations}
  \item{obatch}{An \code{oligoBatch} object.}
}
\keyword{manip}% at least one, from doc/KEYWORDS
