\name{initialize-methods}
\docType{methods}
\alias{initialize-methods}
\alias{initialize,ANY-method}
\alias{initialize,traceable-method}
\alias{initialize,signature-method}
\alias{initialize,environment-method}
\alias{initialize,AnnotatedDataFrame-method}
\alias{initialize,eSet-method}
\alias{initialize,ExpressionSet-method}
\alias{initialize,SnpSet-method}
\alias{initialize,FeatureSet-method}
\title{Methods for Class Initialization}
\description{
  Methods for function \code{initialize} in Package `oligo'.
}
\section{Methods}{
\describe{

\item{.Object = "ANY"}{anything}

\item{.Object = "traceable"}{traceable}

\item{.Object = "signature"}{signature}

\item{.Object = "environment"}{environment}

\item{.Object = "AnnotatedDataFrame"}{AnnotatedDataFrame}

\item{.Object = "eSet"}{eSet}

\item{.Object = "ExpressionSet"}{ExpressionSet}

\item{.Object = "SnpSet"}{SnpSet}

\item{.Object = "FeatureSet"}{FeatureSet object.}
}}
\keyword{methods}
