\name{getChrXIndex}
\alias{getChrXIndex}
\alias{getGenotypeRegionParams}
\alias{getSnpFragmentLength}
\alias{normalizeToSample}
\alias{initialize,DBPDInfo-method}
\alias{initialize,FeatureSet-method}
\alias{initialize,SnpCallSet-method}
\alias{initialize,SnpCopyNumberSet-method}
\alias{initialize,SnpQSet-method}
\alias{show,PDInfo-method}
\alias{show,platformDesign-method}
\alias{length,FeatureSet-method}
\alias{featureInfo}
\alias{featureInfo,platformDesign-method}
\alias{show,PDInfo-method}
\alias{show,platformDesign-method}
\alias{names,platformDesign-method}
\alias{platformDesignName}
\alias{platformDesignName,FeatureSet-method}
\alias{geneNames,AffySNPPDInfo-method}
\alias{geneNames,platformDesign-method}
\alias{geneNames,FeatureSet-method}
\alias{$,platformDesign-method}
\alias{[,FeatureSet-method}
\alias{db}
\alias{db,DBPDInfo-method}
\alias{exprs,FeatureSet-method}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internals}
\description{
}
\usage{
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\details{
}
\value{
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}

\keyword{manip}
