\name{kind}
\alias{kind}
\alias{kind,platformDesign-method}
\alias{kind,AffySNPPDInfo-method}
\alias{kind,AffySNPCNVPDInfo-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Array type}
\description{
  Retrieves the array type.
}
\usage{
kind(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{FeatureSet} object}
}
\details{
}
\value{
  String: "Expression", "Exon", "SNP" or "Tiling"
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}

\keyword{manip}
