\name{list.celfiles}
\alias{list.celfiles}
\alias{list.xysfiles}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{List CEL/XYS files}
\description{
  Lists the CEL/XYS files.
}
\usage{
list.celfiles(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{parameters to be passed to \code{list.files}}
}
\value{
  Character vector with the filenames.
}
\examples{
list.xysfiles()
list.celfiles()
}
\keyword{file}

