\name{list.xysfiles}
\alias{list.xysfiles}
\title{List XYS files}
\description{
  Lists the XYS files.
}
\usage{
list.xysfiles(...)
}
\arguments{
  \item{\dots}{parameters to be passed to \code{\link{list.files}}}
}
\value{
  Character vector with the filenames.
}
\details{
  The functions interface \code{\link{list.files}} and the user is asked
  to check that function for further details.
}
\examples{
list.xysfiles()
}
\seealso{\code{\link{list.files}}}
\keyword{file}

