\name{manufacturer}
\alias{manufacturer}
\alias{manufacturer,FeatureSet-method}
\alias{manufacturer,PDInfo-method}
\alias{manufacturer<-}
\alias{manufacturer<-,FeatureSet-method}

\title{Accessor to the manufacturer slot.}
\description{
  Accessor to the manufacturer slot.
}
\usage{
manufacturer(object)
manufacturer(object)<-value
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{PDInfo} or \code{FeatureSet} object}
  \item{value}{string, eg. "affymetrix" or "nimblegen"}
}
\details{
}
\value{
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}

\keyword{manip}
