\name{nProbes}
\alias{nProbes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Number of probes on the array}
\description{
  The number of probes on the array.
}
\usage{
nProbes(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{a \code{platformDesign} object.}
}
\keyword{methods}% at least one, from doc/KEYWORDS
