\name{PDInfo-class}
\docType{class}
\alias{PDInfo-class}
\alias{SNPPDInfo-class}

\title{Classes "PDInfo" and "SNPPDInfo"}
\description{Containers for chip information}
\section{Objects from the Class}{
}
\section{Slots}{
	 \describe{
    \item{\code{manufacturer}:}{Object of class \code{"character"}  }
    \item{\code{genomebuild}:}{Object of class \code{"character"}  }
    \item{\code{geometry}:}{Object of class \code{"integer"} with length
    2 (rows x columns)}
  }
}
\section{Methods}{
No methods defined with class "PDInfo" in the signature.
}
\references{}
\author{}
\note{}

\seealso{
}
\examples{
##---- Should be DIRECTLY executable !! ----
}
\keyword{classes}
