\name{listFeatureFields}
\alias{listFeatureFields}
\alias{listFeatureSetFields}
\alias{listFeatureFields,PDInfo-method}
\alias{listFeatureSetFields,PDInfo-method}
\alias{listFeatureFields,AffySNPPDInfo-method}
\alias{listFeatureSetFields,AffySNPPDInfo-method}

\title{List fields in the (pm/mm)feature table}
\description{
  Lists the fields in the (pm/mm)feature or featureSet table of the PDInfo object.
}
\usage{
listFeatureFields(object)
listFeatureSetFields(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{PDInfo} object}
}
\details{
}
\value{
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{manip}
