\name{platform}
\alias{platform}
\alias{platform<-}
\alias{platform,FeatureSet-method}
\alias{platform<-,FeatureSet-method}

\title{Retrieve/replace platform information}
\description{
  Retrieve or replace the platform information.
}
\usage{
platform(object)
platform(object)<-value
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{FeatureSet} object.}
  \item{value}{string}
}
\details{
}
\value{
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}

\keyword{manip}
