\name{plotDensity}
\alias{plotDensity}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Plot density of intensities.}
\description{
  Plot density of intensities.
}
\usage{
plotDensity(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{FeatureSet} object}
  \item{\dots}{parameters to be passed to plot.}
}
\details{
}
\value{
}
\references{}
\author{}
\note{}
\seealso{\code{\link{hist}}}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{hplot}
