\name{plotM-methods}
\docType{methods}
\alias{plotM}
\alias{plotM-methods}
\alias{plotM,SnpQSet,character-method}
\alias{plotM,SnpQSet,integer-method}
\alias{plotM,SnpQSet,numeric-method}
\alias{plotM,TilingQSet,missing-method}
\title{Methods for Log-Ratio plotting}
\description{
  The \code{plotM} methods are meant to plot log-ratios for different
  classes of data.
}
\section{Methods}{
  \describe{
    \item{object = "SnpQSet", i = "character"}{Plot log-ratio for SNP data for sample i.}
    \item{object = "SnpQSet", i = "integer"}{Plot log-ratio for SNP data for sample i.}
    \item{object = "SnpQSet", i = "numeric"}{Plot log-ratio for SNP data for sample i.}
    \item{object = "TilingQSet", i = "missing"}{Plot log-ratio for Tiling data for sample i.}
  }
}
%% Add example
\keyword{methods}

