\name{pmAllele}
\alias{pmAllele}
\alias{pmAllele,SnpFeatureSet-method}
\alias{pmAllele,PDInfo-method}
\alias{pmAllele,AffySNPPDInfo-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Access the allele information for PM probes.}
\description{
  Accessor to the allelic information for PM probes.
}
\usage{
pmAllele(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{SnpFeatureSet} or \code{PDInfo} object.}
}
\details{
}
\value{
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{manip}
