\name{pmFragmentLength}
\alias{pmFragmentLength}
\alias{pmFragmentLength,AffySNPPDInfo-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Access the fragment length for PM probes.}
\description{
  Accessor to the fragment length for PM probes.
}
\usage{
pmFragmentLength(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{PDInfo} object.}
}
\keyword{manip}
