\name{pmStrand}
\alias{pmStrand}
\alias{pmStrand,AffySNPPDInfo-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Accessor to the strand information}
\description{
  Returns the strand information on SNP arrays for PM probes (0 - sense / 1 - antisense).
}
\usage{
pmStrand(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{AffySNPPDInfo} object}
}

\keyword{manip}
