\name{probeNames-methods}
\docType{methods}
\alias{probeNames-methods}
\alias{probeNames,platformDesign,characterOrNULL-method}
\alias{probeNames,FeatureSet,characterOrNULL-method}
\title{Probe Names}
\description{
  Extracting the probe names.
}
\section{Methods}{
\describe{

\item{object = "platformDesign", subset = "characterOrNULL"}{extracts the
  probe names for a platformDesign object. \code{subset} is not
  implemented yet.}

\item{object = "FeatureSet", subset = "characterOrNULL"}{extracts the probe
  names for a \code{FeatureSet} object. \code{subset} is not implemented
yet.}
}}
\keyword{methods}
