\name{runDate}
\alias{runDate}
\alias{runDate-methods}
\alias{runDate,FeatureSet-method}
\title{
  Date of scan
}
\description{
  Retrieves date information in CEL/XYS files.
}
\usage{
runDate(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{'FeatureSet' object.}
}
\keyword{ classif }
