\name{snprma}
\alias{snprma}
\title{Preprocessing SNP Arrays}
\description{
  This function preprocess SNP arrays.
}
\usage{
snprma(object, verbose = TRUE, normalizeToHapmap = TRUE)
}
\arguments{
  \item{object}{\code{SnpFeatureSet} object}
  \item{verbose}{Verbosity flag. \code{logical}}
  \item{normalizeToHapmap}{internal}
}
\value{
  A \code{SnpQSet} object.
}
%% ADD EXAMPLE
\keyword{manip}
