\name{boxplot}
\alias{boxplot}
\alias{boxplot,FeatureSet-method}
\alias{boxplot,ExpressionSet-method}
\title{Boxplot}
\description{
  Boxplot for observed (log-)intensities in a FeatureSet-like object
  (ExpressionFeatureSet, ExonFeatureSet, SnpFeatureSet, TilingFeatureSet).
}
\usage{
boxplot(x, ...)
}
%- maybe also 'usage' for other objects documented here.

\arguments{
  \item{x}{a \code{FeatureSet}-like object}
  \item{\dots}{arguments to be passed to plot}
}
\details{
}
\value{
}
\references{}
\author{}
\note{}
\seealso{\code{\link{hist}}, \code{\link{image}}}
\examples{
}

\keyword{hplot}
