\name{getPlatformDesign}
\alias{getPlatformDesign}
\alias{getPD}

\title{Platform Design accessor}
\description{
  Accessor to the platform design object.
}
\usage{
getPlatformDesign(object)
getPD(object)
}

\arguments{
  \item{object}{oligoBatch object.}
}
\value{
  The function returns the platform design object associated to the
  oligoBatch object.
}
\examples{
## Do not  run:
## oBatch <- read.celfiles(list.celfiles())
## getPD(oBatch)
}
\keyword{methods}% at least one, from doc/KEYWORDS
