\name{getPlatformDesign}
\alias{getPlatformDesign}
\alias{getPD}
\alias{getPlatformDesign,FeatureSet-method}

\title{Retrieve Platform Design object}
\description{
  Retrieve platform design object.
}
\usage{
getPlatformDesign(object)
getPD(object)
}
\arguments{
  \item{object}{\code{FeatureSet} object}
}
\details{
    Retrieve platform design object.	
}
\value{
  \code{platformDesign} or \code{PDInfo} object.
}
\keyword{manip}

