\name{hist}
\alias{hist}
\alias{hist,FeatureSet-method}
\alias{hist,ExpressionSet-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Density estimate}
\description{
  Plot the density estimates for each sample
}
\usage{
hist(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{\code{FeatureSet} object}
  \item{\dots}{arguments to be passed to \code{lines}}
}
\details{
}
\value{
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{hplot}

