\name{hist}
\alias{hist}
\alias{hist,FeatureSet-method}
\alias{hist,ExpressionSet-method}
\title{Density estimate}
\description{
  Plot the density estimates for each sample
}
\usage{
hist(x, ...)
}
\arguments{
  \item{x}{\code{FeatureSet} object}
  \item{\dots}{arguments to be passed to \code{lines}}
}
%% Add examples
\keyword{hplot}

