\name{image}
\alias{image}
\alias{image,FeatureSet-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Display a Color Image}
\description{
  Produces an image (\code{graphics::image}) for each sample.
}
\usage{
image(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{\code{FeatureSet} object}
  \item{\dots}{parameters to be passed to \code{plot}}
}
\details{
}
\value{
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}

\keyword{hplot}
