\name{list.celfiles}
\alias{list.celfiles}
\alias{list.xysfiles}
\title{List CEL/XYS files}
\description{
  Lists the CEL/XYS files.
}
\usage{
list.celfiles(...)
}
\arguments{
  \item{\dots}{parameters to be passed to \code{\link{list.files}}}
}
\value{
  Character vector with the filenames.
}
\details{
  The functions interface \code{\link{list.files}} and the user is asked
  to check that function for further details.
}
\examples{
list.xysfiles()
list.celfiles()
}
\seealso{\code{\link{list.files}}}
\keyword{file}

