\name{MAplot-methods}
\docType{methods}
\alias{MAplot}
\alias{MAplot-methods}
\alias{MAplot,FeatureSet-method}
\title{MA plots}
\description{
  Create MA plots using a reference array (based on medians).
}
\section{Methods}{
\describe{

\item{object = "FeatureSet"}{ExpressionFeatureSet}
}}
\keyword{methods}

