\name{MAplot-methods}
\docType{methods}
\alias{MAplot}
\alias{MAplot-methods}
\alias{MAplot,FeatureSet-method}
\alias{MAplot,PLMset-method}
\title{MA plots}
\description{
  Create MA plots using a reference array (if one channel) or using
  channel2 as reference (if two channel).
}
\section{Methods}{
  \describe{
    \item{object = "FeatureSet"}{ExpressionFeatureSet}
  }
}

\keyword{methods}

