\name{names}
\alias{names}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Platform Design variables}
\description{
  A function to list the variables available in the platform design object.
}
\usage{
names(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{platform design object}
}
\value{
  The function returns a list of variables available in the platform
  design object.
}
\seealso{\code{\link{getPD}}}
\examples{
## DO NOT RUN
## FeatureSetObject <- read.celfiles(list.celfiles())
## names(getPD(FeatureSetObject))
}
\keyword{methods}% at least one, from doc/KEYWORDS
