\name{nProbes}
\alias{nProbes}
\alias{nProbes,platformDesign-method}
\alias{nProbes,DBPDInfo-method}
%% \alias{nProbes,AffySNPPDInfo-method}
%% \alias{nProbes,ExpressionPDInfo-method}
%% \alias{nProbes,TilingPDInfo-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Count number of probes}
\description{
  Counts the number of probes in a \code{platformDesign} object.
}
\usage{
nProbes(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{platformDesign} object}
}
\keyword{manip}
