\name{paCalls}
\alias{paCalls}
\alias{paCalls,ExonFeatureSet-method}
\title{Methods for P/A Calls}
\description{
  Methods for Present/Absent Calls are meant to provide
  means of assessing whether or not each of the (PM) intensities are
  compatible with observations generated by background probes.
}
\usage{
paCalls(object, method, verbose = TRUE)
}
\arguments{
  \item{object}{\code{ExonFeatureSet} object.}
  \item{method}{String defining what method to use. Currently "dabg".}
  \item{verbose}{Logical flag for verbosity.}
}
\value{
  A matrix (of dimension dim(PM))  with p-values for P/A Calls.
}
\references{
  Clark et al. Discovery of tissue-specific exons using comprehensive human exon microarrays. Genome Biol (2007) vol. 8 (4) pp. R64
}
\author{
  Benilton Carvalho
}
\examples{
if (require(oligoData) & require(pd.huex.1.0.st.v2)){
  data(affyExonFS)
  ## Get only 2 samples for example
  affyExonFS = affyExonFS[, 1:2]
  dabgP = paCalls(affyExonFS)
  head(dabgP)
}
}
\keyword{manip}
