\name{listFeatureFields}
\alias{listFeatureFields}
\alias{listFeatureSetFields}
\alias{listFeatureFields,PDInfo-method}
\alias{listFeatureSetFields,PDInfo-method}
\alias{listFeatureFields,AffySNPPDInfo-method}
\alias{listFeatureSetFields,AffySNPPDInfo-method}
\alias{geometry}
\alias{geometry,PDInfo-method}
\alias{geometry,DBPDInfo-method}

\title{List fields in the (pm/mm)feature table}
\description{
  Lists the fields in the (pm/mm)feature or featureSet table of the PDInfo object.
}
\usage{
listFeatureFields(object)
listFeatureSetFields(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{PDInfo} object}
}
\keyword{manip}
