\name{platformDesignName}
\alias{platformDesignName}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{The platform design object name.}
\description{
  The name of \code{platformDesign} object associated to the \code{oligoBatch} object.
}
\usage{
platformDesignName(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{An \code{oligoBatch} object.}
}
\value{
  A string with the name of the PDEnv object name.
}
\keyword{manip}% at least one, from doc/KEYWORDS
