\name{pmAllele}
\alias{pmAllele}
\alias{pmAllele,AffySNPPDInfo-method}
\alias{pmAllele,SnpFeatureSet-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Access the allele information for PM probes.}
\description{
  Accessor to the allelic information for PM probes.
}
\usage{
pmAllele(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{SnpFeatureSet} or \code{PDInfo} object.}
}
\keyword{manip}
