\name{pmPosition}
\alias{pmPosition}
\alias{pmPosition,AffySNPPDInfo-method}
\alias{pmPosition,TilingFeatureSet-method}
\alias{position}
\alias{position,TilingFeatureSet-method}
\alias{pmPosition,ExpressionPDInfo-method}
\alias{pmPosition,TilingPDInfo-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Accessor to position information}
\description{
  For tiling arrays, this will return the genomic position for the
  (PM) probes. For SNP arrays, offset information will be returned.
}
\usage{
pmPosition(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{AffySNPPDInfo} or \code{TilingFeatureSet} object}
}
\details{
}
\value{
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}

\keyword{manip}
