\name{pmStrand}
\alias{pmStrand}
\alias{pmStrand,AffySNPPDInfo-method}
\alias{pmStrand,TilingFeatureSet-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Accessor to the strand information}
\description{
  Returns the strand information for PM probes (0 - sense / 1 - antisense).
}
\usage{
pmStrand(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{AffySNPPDInfo} or \code{TilingFeatureSet} object}
}

\keyword{manip}
