\name{probeNames}
\alias{probeNames}
\alias{probeNames,FeatureSet-method}
\alias{probeNames,AffySNPPDInfo-method}
\alias{probeNames,platformDesign-method}
\alias{probeNames,ExpressionPDInfo-method}
\alias{probeNames,TilingPDInfo-method}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{Accessor to feature names}
\description{
  Accessor to PM feature names.
}
\usage{
probeNames(object, subset = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{FeatureSet}, \code{AffySNPPDInfo} or
    \code{platformDesign} object.}
  \item{subset}{not implemented yet.}
}
\details{
}
\value{
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}

\keyword{manip}
