\name{probeNames}
\alias{probeNames}
\alias{probeNames,FeatureSet-method}
\alias{probeNames,ExonFeatureSet-method}
\alias{probeNames,GeneFeatureSet-method}
\alias{probeNames,DBPDInfo-method}
\title{Accessor to feature names}
\description{
  Accessor to PM feature names.
}
\usage{
probeNames(object, subset = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{FeatureSet} or \code{DBPDInfo}}
  \item{subset}{not implemented yet.}
}
\keyword{manip}
