\name{sequenceDesignMatrix}
\alias{sequenceDesignMatrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Create design matrix for sequences}
\description{
  Creates design matrix for sequences.
}
\usage{
sequenceDesignMatrix(seqs)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{seqs}{\code{character} vector of 25-mers.}
}
\details{
}
\value{
  Matrix with length(seqs) rows and 75 columns.
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{manip}
