\name{mmSequence}
\alias{mmSequence}
\alias{pmSequence}
\alias{mmSequence,FeatureSet-method}
\alias{pmSequence,FeatureSet-method}
\alias{mmSequence,platformDesign-method}
\alias{pmSequence,platformDesign-method}
\alias{mmSequence,AffySNPPDInfo-method}
\alias{pmSequence,AffySNPPDInfo-method}
\alias{pmSequence,ExpressionPDInfo-method}
\alias{pmSequence,TilingPDInfo-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Probe Sequeces}
\description{
  Accessor to the probe sequences.
}
\usage{
mmSequence(object)
pmSequence(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{FeatureSet}, \code{platformDesign} or
    \code{AffySNPPDInfo} object}
}
\details{
}
\value{
  A character vector with the PM or MM probe sequences.
}

\references{}
\author{}
\note{}
\seealso{}
\examples{
}

\keyword{manip}
