\name{snprma}
\alias{snprma}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Preprocessing SNP Arrays}
\description{
  This function preprocess SNP arrays.
}
\usage{
snprma(oBatch, normalizeToHapmap = TRUE, saveQuant = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{oBatch}{\code{SnpFeatureSet} object}
  \item{normalizeToHapmap}{internal}
  \item{saveQuant}{internal}
}
\details{
}
\value{
  A \code{SnpQSet} object.
}
\references{}
\author{}
\note{}
\seealso{}
\examples{
}

\keyword{manip}
